+++ reactivestate-compose/src/commonMain/kotlin/com/ensody/reactivestate/compose/ReactiveStateComposeLib.kt
@@ -1,4 +1,0 @@
-package com.ensody.reactivestate.compose
-
-/** This simply exists to prevent an empty klib. */
-internal val zzzNonEmpty = true